# Inline low level multiplication and reduction functions