@fjarri 2025-02-22
Actually wait a second, the same effect is achieved by just inlining.